Windows Task Scheduler version
Added bypassing bot checks, adblock, full list of websites and their minimal withdrawal amount at the moment, login and rolling optimization, file logging.
Excluded Twitter codes and price info as being unneccassary as this version is meant to be automatic.

Initial purpose of this changes is to use the bot with Windows Task Scheduler to run it every hour

*Please mind that this is my first commit on github so I may not understand commit workflow fully